+++ src/App.js
@@ -1,60 +1,10 @@
-import React, { Component, PureComponent, createElement } from 'react';
+import React from 'react';
 
-class HeaderTitle extends PureComponent {
-  render() {
-    return <h1>{this.props.title}</h1>;
-  }
-}
+import { Counter } from './components';
 
-const FinalCounter = ({ counter }) => <h2>counter: {counter}</h2>;
-
-const ButtonCounter = ({ handle, text }) =>
-  createElement('button', { onClick: handle }, text);
-
-class App extends Component {
-  constructor() {
-    super();
-    this.state = {
-      counter: 0,
-      title: 'My Counter',
-    };
-  }
-
-  increment = () => {
-    this.setState({
-      counter: this.state.counter + 1,
-    });
-  };
-
-  decrement = () => {
-    this.setState({
-      counter: this.state.counter - 1,
-    });
-  };
-
-  reset = () => {
-    this.setState({
-      counter: 0,
-    });
-  };
-  render() {
-    const { counter, title } = this.state;
-    return (
-      <main>
-        <header>
-          <HeaderTitle title={title} />
-        </header>
-        <section>
-          <FinalCounter counter={counter} />
-          <div>
-            <ButtonCounter handle={this.increment} text={'Increment!'} />
-            <ButtonCounter handle={this.decrement} text={'Decrement!'} />
-            <ButtonCounter handle={this.reset} text={'Reset!'} />
-          </div>
-        </section>
-      </main>
-    );
-  }
-}
-
+const App = () => (
+  <main>
+    <Counter />
+  </main>
+);
 export default App;
